python_client/pyproject.toml
[project]
name = "python-client"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "vaibhav", email = "vaibhav.229301110@muj.manipal.edu" }
]
requires-python = ">=3.10"
dependencies = [
    "grpcio>=1.78.1",
    "grpcio-tools>=1.78.1",
    "pyqtgraph>=0.14.0",
    "pyside6>=6.10.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["ui"]
